Controller instead of Game
The main controlling class is now called Controller instead of Game. A
new instance of Game will now be created whenever a new Game is started.
Game will controll this game, manage the score and lives and create new
Items. An instancte of Controller is created at the start of the
Aplication and will always stay active. It mainly manages the instances
of all other classes and lets them communicate.